Add several new equipment

+++ src/main/java/me/neoblade298/neorogue/equipment/armor/Footpads.java
@@ -1,18 +1,17 @@
 package me.neoblade298.neorogue.equipment.armor;
-import me.neoblade298.neorogue.equipment.SessionEquipment;
-
 import org.bukkit.Material;
 
 import me.neoblade298.neorogue.DescUtil;
 import me.neoblade298.neorogue.equipment.Equipment;
 import me.neoblade298.neorogue.equipment.Rarity;
+import me.neoblade298.neorogue.equipment.SessionEquipment;
 import me.neoblade298.neorogue.session.fight.PlayerFightData;
 import me.neoblade298.neorogue.session.fight.trigger.Trigger;
 import me.neoblade298.neorogue.session.fight.trigger.TriggerResult;
 
 public class Footpads extends Equipment {
 	private static final String ID = "Footpads";
-	private double stamina;
+	private int stamina;
 	
 	public Footpads(boolean isUpgraded) {
 		super(ID, "Footpads", isUpgraded, Rarity.COMMON, new EquipmentClass[] { EquipmentClass.WARRIOR, EquipmentClass.THIEF },
